bug: hack: add GetClusterProvider() function
/kind bug
/area service-mesh

Add small hack that tries to determine the hyperscaler based on the first provider node.
When provider is not known, returns ErrUnknownProvider error which needs to be properly handled.

Gardener clusters are usually backed by the single hyperscaler, so all nodes in the K8s cluster contain the same prefix for providerID.

The only thing it may fail is when user adds nodes manually by hand and they have different providerIDs, which is very unlikely in gardener cluster.