bug: fix fetching main image from registy
/kind bug
/area ci

There is a problem with properly assuming which tag should ba taken as the latest, because docker registry returns all tags for image only in alphabetical order. SHAs are not alphabetical, and sorting them is tedious task.

Because we build images and tag them with full commit SHA, instead of trying to fetch and sort all tags, grab latest HEAD of git repo and check if the tag exists on the repo. If it exists, update config. Otherwise, skip updating config update